Introduce `ChatChannelListSearchVC` abstract class to reuse searching logic
